Split handling of kubernetes.io Anno’s + Labels
The official Kubernetes explicitly prevents the use of annotations and labels that have “kubernetes.io” as part of the key name. This is because these are often added server side, by the Kubernetes API and would cause a diff on the Terraform side which has no knowledge of the key.

This commit updates the Terraform logic to allow kubernetes.io keys, but on Read, strip any that aren’t found in the Terraform config / schema. This way a diff is not triggered.