kubernetes/resource_kubernetes_persistent_volume: storage_class_name is now computed
Set 'storage_class_name' to be computed for the persistent volume
resource. The persistent volume claim resource does the same thing, so
this should mimic the behaviour in persistent volume claim.

Having one of the storage_class_name's be computed and the other not
causes issues when trying to create a persistent volume and persistent
volume claim, because if there is a StorageClass already created in
Kubernetes, that will be used for the persistent volume claim
storage_class_name, but the persistent volume will be created with an
empty storage_class_name and this fails because Kubernetes validates
that both storage_class_name is the same for both the persistent volume
and persistent volume claim.