Allow max_unavailable and max_surge to be 0
This allows for the valid case where either max_surge or max_unavailable
is 0 but the other is non-0.

At least one of these must be non-0, which means this change will allow
both values to be set to 0 which could cause issues. Suggest some way of
validating a resource (potentially add some kind of validation function
parameter to schema.Resource?) to ensure dependencies between different
parameters are being met.