accept PATH-style kubeconfig
The env var KUBECONFIG may be set in the style of $PATH, i.e., mutliple
files divided by a colon, so that the actual configuration may be spread
out over multiple files. THis may be desirable so that the non-secret
cluster configuration may be checked into a source code repo leaving
actual credentials separate.