Support in-cluster configuration with service accounts (#497)
Prior to this commit, the provider supports authenticating through
a local `kubeconfig` or through various configuration provided directly
to the provider or from the environment.

The provider however, did not authenticate using a service account
when running inside a Kubernetes pod. This commit provides support
for loading in-cluster configuration so that the provider can
be used inside a Kubernetes pod (with a service account) without
much additional configuration.

The use-case for this is running Terraform via Atlantis inside
a Kubernetes pod. But, outside of Atlantis, there may be cases where
we want to run some Terraform automation inside a pod (e.g. GitLab runner).