Add wait_for_rollout attribute to kubernetes_deployment.
This controls whether the provider will wait for a create or update
operation to be completely applied on the server side. For some
situations this is desirable behaviour, but in other situations where a
gradual rollout using the builtin Kubernetes controls can take hours
this is undesirable.

It is set to true by default (matching current behaviour).

This should be done for kubernetes_stateful_set too, but it wasn't clear
how to wait for a StatefulSet change to complete.

Updates #605.