New datasources: kubernetes_namespace & kubernetes_all_namespaces (#701)
This change introduces a couple of new data sources for reading Kubernetes Namespace resources.

`kubernetes_namespace` is a standard data source that exposes the attributes of a given namespace.

`kubernetes_all_namespaces` is an enumerating data source that lists all available namespaces within a cluster. It can be used to test for the presence of a specific namespace or for applying a resource to all namespaces.