Allow in-place edits of PVC's storage request
The Kubernetes API allows in-place edits of a PVC's spec.resources.requests.storage field.
This change updates the PVC resource to allow the same.

To avoid surfacing this Kubernetes API error to users, the PVC will be recreated if the
specified storage value is less than the previously specified value:

`spec.resources.requests.storage: field can not be less than previous value`

This saves users the trouble of deleting and recreating the PVC themselves.