bugfix: introduce filter for flattenPodSpec to control tolerations
If you start a pod natively, k8s may add tolerations automatically that
you did not define. There is already a filter in the flattenPodSpec that
filters these out, but if you start it as part of a
deployment/replicaset/daemonset, you might want to use similar
tolerations and don't want to get in a perpetual diff (where state
refresh ignores your toleration in deployments just like it would with
native pods).

This is not a silver bullet, but unblocked me at my work, in the future
maybe a more robust solution (e.g. there is not tellign that you will
use this prefix of tolerations in natively started pods, making the
problem I have with deployments appear there as well)