Fix regression tests and remove self_link from metadata (#1110)
The regression tests which run StateUpgraders on the optional `resources {}` attribute have stopped working. This commit adds the `resources {}` attribute explicitly to each regression test, which allows the StateUpgrader to run. (This is only needed during acceptance tests, since the StateUpgrader is able to run during a manual terraform refresh or terraform apply).

This also removes SelfLink from metadata, since that field is no longer populated as of Kubernetes version 1.20. (Otherwise, acceptance tests fail when running on 1.20).