Clarify docs for role_binding.subject.api_group
This field was listed as "Required", but is only required for User or Group subjects, not where type = ServiceAccount. For ServiceAccount subjects, setting this value results in an error when you try to apply the change: `RoleBinding.rbac.authorization.k8s.io "{name}" is invalid: subjects[0].apiGroup: Unsupported value: "rbac.authorization.k8s.io": supported values: ""`  

It's a tiny change, but I think this is clearer now?