Add aliases for v1 resources (#1484)
This commit adds a version suffix to each resource. Most resources are v1, with the exception of PodSecurityPolicy and HorizontalPodAutoscaler which also have v1beta1 and v2beta2 respectively.

This is so Terrafom has an explicitly versioned schema for each resource version, rather than having to include attributes from different API versions in a single schema and figure out which API version needs to be used. This will make the provider easier to maintain, and subsequently to automate the code generation of resource schemas for each API version.