Remove ArgoCD operation field to prevent perpetual diff
When using the `kubernetes_manifest` resource to manage an ArgoCD
Application, it is possible for the status of Argo's sync operation to
show up in the manifest when applying, but not during planning. This
shows up under the top-level `operation` field, which is only present
when an Argo sync operation is underway, potentially triggered by the
Terraform apply.

This commit simply drops the `operation` field from manifests returned
by the K8s API and thusly prevents it from being considered when
diffing.