fix: Resolve panic with `ElementKeyInt(0): can't use tftypes`
This commit updates the `manifest/morph/scaffold.go` logic in order to
resolve a panic encountered when working with `Tuple` types.

Also adds a new unit test to cover this case.

Disclaimer: This fix and the coresponding test case was generated by
Gemini, and has been reviewed and tested by myself.

Fixes: #2778 #2754 plus probably others...